% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{adapt_data_types}
\alias{adapt_data_types}
\title{Adapt data types for database insertion}
\usage{
adapt_data_types(data, debug = FALSE)
}
\arguments{
\item{data}{Data to adapt}

\item{debug}{Enable debug messages (optional, default FALSE)}
}
\value{
Adapted data
}
\description{
Adapt data types for database insertion
}
\examples{
data <- data.frame(
  timestamp = Sys.time() + 1:5,
  date = Sys.Date() + 1:5,
  value = rnorm(5)
)
adapted_data <- adapt_data_types(data)
}
