% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{clean_data_for_primary_keys}
\alias{clean_data_for_primary_keys}
\title{Clean data for primary keys}
\usage{
clean_data_for_primary_keys(data, primary_keys, debug = FALSE)
}
\arguments{
\item{data}{Data to clean}

\item{primary_keys}{Vector of primary key column names}

\item{debug}{Enable debug messages (optional, default FALSE)}
}
\value{
Cleaned data
}
\description{
Clean data for primary keys
}
\examples{
data <- data.frame(
  id = c(1, 2, NA, 4, 5),
  value = rnorm(5)
)
cleaned_data <- clean_data_for_primary_keys(data, "id")
}
