% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_timeseries}
\alias{clean_timeseries}
\title{Clean and transform data with dplyr and tidyr}
\usage{
clean_timeseries(data, time_col, value_cols, pivot = FALSE)
}
\arguments{
\item{data}{A data frame containing time series data}

\item{time_col}{The column containing timestamp data}

\item{value_cols}{Vector of value column names to process}

\item{pivot}{Whether to pivot the data to long format (default: FALSE)}
}
\value{
Processed data frame
}
\description{
Cleans and transforms time series data using dplyr and tidyr functions.
}
\examples{
\dontrun{
# Get data from database
data <- get_query("SELECT time, temp, humidity, pressure FROM sensor_data LIMIT 100")

# Clean and transform data
clean_data <- clean_timeseries(
  data,
  time_col = "time",
  value_cols = c("temp", "humidity", "pressure"),
  pivot = TRUE
)
}
}
