% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to database}
\usage{
connect_to_db(config)
}
\arguments{
\item{config}{Connection configuration}
}
\value{
Database connection object
}
\description{
Connect to database
}
\examples{
\dontrun{
conn <- connect_to_db(config)
}
}
