% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{debug_connection}
\alias{debug_connection}
\title{Debug connection}
\usage{
debug_connection(config)
}
\arguments{
\item{config}{Connection configuration}
}
\value{
List with connection information
}
\description{
Debug connection
}
\examples{
\dontrun{
debug_connection(config)
}
}
