% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{determine_column_types}
\alias{determine_column_types}
\title{Determine column types for table creation}
\usage{
determine_column_types(data, debug = FALSE)
}
\arguments{
\item{data}{Data to analyze}

\item{debug}{Enable debug messages (optional, default FALSE)}
}
\value{
Vector of SQL types
}
\description{
Determine column types for table creation
}
\examples{
data <- data.frame(
  timestamp = Sys.time(),
  date = Sys.Date(),
  number = 123.45,
  flag = TRUE,
  text = "example"
)
col_types <- determine_column_types(data)
}
