% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_csv}
\alias{import_csv}
\title{Read data from CSV file}
\usage{
import_csv(file, stringsAsFactors = FALSE)
}
\arguments{
\item{file}{Path to CSV file}

\item{stringsAsFactors}{Whether to convert strings to factors (default: FALSE)}
}
\value{
Data frame
}
\description{
Reads data from a CSV file using utils::read.csv
}
\examples{
\dontrun{
# Read data from CSV
data <- import_csv("sensor_data.csv")

# View structure
str(data)
}
}
