% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_json}
\alias{parse_json}
\title{Convert JSON data to R objects}
\usage{
parse_json(json_string)
}
\arguments{
\item{json_string}{A JSON string}
}
\value{
Parsed R object
}
\description{
Utility function to convert JSON data from TimescaleDB to R objects.
}
\examples{
\dontrun{
# Get JSON data from database
result <- get_query("SELECT data FROM json_data LIMIT 1")
json_str <- result$data[1]

# Parse JSON
parsed_data <- parse_json(json_str)
str(parsed_data)
}
}
